requirement.txt
streamlit
ultralytics
pydantic[email]
python-jose[cryptography]
passlib[argon2]
uvicorn 
sqlalchemy 
numpy
pandas
Pillow
fastapi
python-dotenv